# OEUPB-Docs/specs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OE UPB - Observatorio de Egresados",
    "description": "API RESTful para el procesamiento de egresados UPB.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/directorio/tabla": {
      "get": {
        "tags": [
          "Directorio"
        ],
        "summary": "Obtener Directorio",
        "operationId": "obtener_directorio_api_directorio_tabla_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "programa",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Programa"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectorioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/directorio/programas": {
      "get": {
        "tags": [
          "Directorio"
        ],
        "summary": "Obtener Programas Unicos",
        "operationId": "obtener_programas_unicos_api_directorio_programas_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array",
                  "title": "Response Obtener Programas Unicos Api Directorio Programas Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/directorio/perfil/{documento}": {
      "get": {
        "tags": [
          "Directorio"
        ],
        "summary": "Obtener Perfil Egresado",
        "operationId": "obtener_perfil_egresado_api_directorio_perfil__documento__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "documento",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Documento"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfilEgresadoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/directorio/egresados": {
      "post": {
        "tags": [
          "Directorio"
        ],
        "summary": "Crear Egresado",
        "operationId": "crear_egresado_api_directorio_egresados_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EgresadoManualRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperacionEgresadoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/directorio/egresados/{documento}": {
      "patch": {
        "tags": [
          "Directorio"
        ],
        "summary": "Editar Egresado",
        "operationId": "editar_egresado_api_directorio_egresados__documento__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "documento",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Documento"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EgresadoManualUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperacionEgresadoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Directorio"
        ],
        "summary": "Eliminar Egresado",
        "operationId": "eliminar_egresado_api_directorio_egresados__documento__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "documento",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Documento"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EliminarEgresadoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperacionEgresadoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/directorio/exportar.xlsx": {
      "get": {
        "tags": [
          "Directorio"
        ],
        "summary": "Exportar Directorio",
        "operationId": "exportar_directorio_api_directorio_exportar_xlsx_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "programa",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Programa"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Directorio Excel",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DirectorioItem": {
        "properties": {
          "documento": {
            "type": "string",
            "title": "Documento"
          },
          "nombre_completo": {
            "type": "string",
            "title": "Nombre Completo"
          },
          "programa": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Programa"
          },
          "fecha_grado": {
            "type": "string",
            "title": "Fecha Grado"
          },
          "encuestas": {
            "type": "string",
            "title": "Encuestas"
          }
        },
        "type": "object",
        "required": [
          "documento",
          "nombre_completo",
          "programa",
          "fecha_grado",
          "encuestas"
        ],
        "title": "DirectorioItem"
      },
      "DirectorioResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/DirectorioItem"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "limit",
          "data"
        ],
        "title": "DirectorioResponse"
      },
      "EgresadoManualRequest": {
        "properties": {
          "numero_documento": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Numero Documento"
          },
          "primer_nombre": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Primer Nombre"
          },
          "primer_apellido": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Primer Apellido"
          },
          "programa": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Programa"
          },
          "fecha_grado": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Grado"
          },
          "motivo": {
            "type": "string",
            "maxLength": 500,
            "minLength": 5,
            "title": "Motivo"
          }
        },
        "type": "object",
        "required": [
          "numero_documento",
          "primer_nombre",
          "programa",
          "motivo"
        ],
        "title": "EgresadoManualRequest"
      },
      "EgresadoManualUpdate": {
        "properties": {
          "primer_nombre": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Primer Nombre"
          },
          "primer_apellido": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Primer Apellido"
          },
          "programa": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Programa"
          },
          "fecha_grado": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Grado"
          },
          "motivo": {
            "type": "string",
            "maxLength": 500,
            "minLength": 5,
            "title": "Motivo"
          }
        },
        "type": "object",
        "required": [
          "motivo"
        ],
        "title": "EgresadoManualUpdate"
      },
...
      "EliminarEgresadoRequest": {
        "properties": {
          "motivo": {
            "type": "string",
            "maxLength": 500,
            "minLength": 5,
            "title": "Motivo"
          }
        },
        "type": "object",
        "required": [
          "motivo"
        ],
        "title": "EliminarEgresadoRequest"
      },
      "EncuestaPerfilResponse": {
        "properties": {
          "momento": {
            "type": "integer",
            "title": "Momento"
          },
          "anio": {
            "type": "integer",
            "title": "Anio"
          },
          "empleabilidad": {
            "title": "Empleabilidad"
          },
          "salario": {
            "title": "Salario"
          },
          "respuestas_completas": {
            "additionalProperties": true,
            "type": "object",
            "title": "Respuestas Completas"
          }
        },
        "type": "object",
        "required": [
          "momento",
          "anio",
          "empleabilidad",
          "salario",
          "respuestas_completas"
        ],
        "title": "EncuestaPerfilResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperacionEgresadoResponse": {
        "properties": {
          "documento": {
            "type": "string",
            "title": "Documento"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          }
        },
        "type": "object",
        "required": [
          "documento",
          "estado"
        ],
        "title": "OperacionEgresadoResponse"
      },
      "PerfilEgresadoResponse": {
        "properties": {
          "documento": {
            "type": "string",
            "title": "Documento"
          },
          "nombre_completo": {
            "type": "string",
            "title": "Nombre Completo"
          },
          "programa": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Programa"
          },
          "fecha_grado": {
            "type": "string",
            "title": "Fecha Grado"
          },
          "encuestas": {
            "items": {
              "$ref": "#/components/schemas/EncuestaPerfilResponse"
            },
            "type": "array",
            "title": "Encuestas"
          }
        },
        "type": "object",
        "required": [
          "documento",
          "nombre_completo",
          "programa",
          "fecha_grado",
          "encuestas"
        ],
        "title": "PerfilEgresadoResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}